Updated to mysqli_* functions, manually addedd install and upgrade folders.
However, branch seems to be a copy from some running system. Hence, it missed some parts (install and upgrade).
Which probably caused problems, as a 'clean' has a config.php that expects version 7, but we only get version 5.
Going to try to revert to production, and make corresponding patches there.